new branch init

+++ bnpy/ioutil/ModelReader.py
@@ -13,7 +13,7 @@ import scipy.io
 import os
 import glob
 
-from .ModelWriter import makePrefixForLap
+from bnpy.ioutil.ModelWriter import makePrefixForLap
 from bnpy.allocmodel import AllocModelConstructorsByName
 from bnpy.obsmodel import ObsModelConstructorsByName
 from bnpy.util import toCArray, as1D, as2D
@@ -199,11 +199,11 @@ def loadDictFromMatfile(matfilepath):
     >>> scipy.io.savemat('Dorig.mat', Dorig, oned_as='row')
     >>> D = loadDictFromMatfile('Dorig.mat')
     >>> D['scalar']
-    array(5)
+    array(5, dtype=int32)
     >>> D['scalar1DN1']
     array(3.14)
     >>> D['arr1DN3']
-    array([1, 2, 3])
+    array([1, 2, 3], dtype=int32)
     '''
     Dtmp = scipy.io.loadmat(matfilepath)
     D = dict([x for x in list(Dtmp.items()) if not x[0].startswith('__')])
